requirements.txt
'django-rest-auth[with_social]'
django
djangorestframework
